### 2025-07-10 #8568 @editorialbot
[PRE REVIEW]: KeemenaPreprocessing.jl: Unicode-Robust Cleaning, Multi-Level Tokenisation & Streaming Offset Bundling for Julia NLP
**Submitting author:** @mantzaris (<a href="http://orcid.org/0000-0002-0026-5725">Alexander V. Mantzaris</a>)
**Repository:** https://github.com/mantzaris/KeemenaPreprocessing.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
