### 2025-08-16 7150c51
now the documentation is improved with more of an introduction for the PreprocessBundle, and some tweeks to the workflows. Also the CodecZlib is now a dependency rather than an extra.
