+++ docs/src/guides/types.md
@@ -1,0 +1,221 @@
+
+# Core Types Introduction
+
+This page gives a practical introduction to the small set of types that underline **KeemenaPreprocessing.jl**. It is aimed at users who want to understand the data model well enough to slice, align, and feed text into downstream models without reading the entire API reference.
+
+---
+
+## Mental model (one minute)
+
+Keemena represents a corpus as a **flat vector of token identifiers** plus **offset vectors** that mark the starts of higher‑level units:
+
+```
+bytes -> characters -> words -> sentences -> paragraphs -> documents
+```
+
+Each segmentation level is packaged as a `LevelBundle` (a `Corpus` + its `Vocabulary`). A `PreprocessBundle` is the top‑level structure that stores all levels, cross‑level alignment tables, a reproducibility header, and optional user payload ('extras').
+
+---
+
+## `PreprocessBundle{ExtraT}`
+
+Top‑level structure returned by `preprocess_corpus` (and the streaming variants).
+
+**Fields**
+
+- `levels :: Dict{Symbol,LevelBundle}` - the available segmentation levels (e.g. `:byte`, `:character`, `:word`, `:sentence`, `:paragraph`, `:document`).
+- `metadata :: PipelineMetadata` - reproducibility header (configuration + on‑disk schema version).
+- `alignments :: Dict{Tuple{Symbol,Symbol},CrossMap}` - cross‑level projections keyed as `(source, destination)`, eg. `(:byte, :word)`
+- `extras :: ExtraT` - optional user payload (feature matrices, raw text, tags, etc.).
+
+**Common operations**
+```julia
+using KeemenaPreprocessing
+
+# Load or create a bundle
+bundle = preprocess_corpus("corpus/*.txt")
+
+# Inspect available levels
+available_levels = collect(keys(bundle.levels))  # eg., [:byte, :character, :word]
+
+# Access a level's vocabulary
+word_vocabulary = bundle.levels[:word].vocabulary
+println("vocabulary size: ", length(word_vocabulary.id_to_token_strings))
+
+# Ensure canonical cross-level maps (:byte -> :character -> :word)
+build_ensure_alignments!(bundle)
+
+# Membership query: which word contains byte index i?
+# (requires the canonical alignments, ensured above)
+byte_index = 1
+word_index_of_byte1 = bundle.alignments[(:byte, :word)].alignment[byte_index]
+```
+
+---
+
+## `LevelBundle`
+
+Pairing of a `Corpus` with its companion `Vocabulary` for **one** segmentation level:
+
+**Fields**
+- `corpus     :: Corpus`
+- `vocabulary :: Vocabulary`
+
+Integrity checks ensure all token identifiers in `corpus.token_ids` are valid for the vocabulary:
+
+**Example**
+```julia
+word_level  = bundle.levels[:word]
+word_corpus = word_level.corpus
+word_vocab  = word_level.vocabulary
+```
+
+---
+
+## `Corpus`
+
+A compact container for all token identifiers at a level and optional offset tables that recover structure:
+
+**Fields**
+- `token_ids          :: Vector{Int}` (always present)
+- `document_offsets   :: Vector{Int}` (always present)
+- `paragraph_offsets  :: Union{Vector{Int},Nothing}`
+- `sentence_offsets   :: Union{Vector{Int},Nothing}`
+- `word_offsets       :: Union{Vector{Int},Nothing}`
+- `character_offsets  :: Union{Vector{Int},Nothing}`
+- `byte_offsets       :: Union{Vector{Int},Nothing}`
+
[…]
