+++ Project.toml
@@ -9,20 +9,21 @@ JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Unicode = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
+CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
 
 [compat]
 Glob = "1"
 JLD2 = "0.5"
 StatsBase = "0.34"
 julia = "1.6"
+CodecZlib = "0.7"
 
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
-CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "Aqua", "BenchmarkTools", "CodecZlib", "Downloads"]
+test = ["Test", "Aqua", "BenchmarkTools", "Downloads"]

[…]
