### 2025-08-20 42211a4
changing the readme instructions on installation
