### 2025-08-29 9e929e7
CompatHelper: bump compat for JLD2 to 0.6, (keep existing compat)
