+++ Project.toml
@@ -4,19 +4,19 @@ authors = ["Alexander V. Mantzaris"]
 version = "0.1.0"
 
 [deps]
+CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
 Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Unicode = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
-CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
 
 [compat]
+CodecZlib = "0.7"
 Glob = "1"
-JLD2 = "0.5"
+JLD2 = "0.5, 0.6"
 StatsBase = "0.34"
 julia = "1.6"
-CodecZlib = "0.7"
 
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
