### 2025-11-04 #9348 @editorialbot
[REVIEW]: KeemenaPreprocessing.jl: Unicode-Robust Cleaning, Multi-Level Tokenisation & Streaming Offset Bundling for Julia NLP
**Submitting author:** @mantzaris (<a href="http://orcid.org/0000-0002-0026-5725">Alexander V. Mantzaris</a>)
**Repository:** https://github.com/mantzaris/KeemenaPreprocessing.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.1
