### 2025-11-15 #9348 @lockwo
Hi @frankier, let me know if you have any questions or concerns with the editorial bot or the review. When you start by running

it sets up the nice streamlined process
