### 2025-11-16 28856f5
CompatHelper: bump compat for JLD2 to 0.6, (keep existing compat)
