### 2025-12-12 f83e0e0
added an example to the docs for the usage of WordTokenizers adapter
