### 2025-12-15 #9348 @lockwo
> Hi [@lockwo](https://github.com/lockwo) and thank you for the reminder. I’ve been down with a high fever for a few days, but I’m already recovering and should be able to resume work tomorrow. Because of this unexpected delay, I’d like to request a short extension of 2–3 days for the review, if possible. I appreciate your understanding. These things are hard to plan for. I’ll likely get back to you tomorrow with any questions or issues related to the process.
>
> Best

Yes this extension is totally reasonable, please take time and prioritize your health. I hope you feel better soon.
