### 2025-12-17 #9348 @atantos
Thank you @frankier and @mantzaris both for the detailed exchange. I think it has clarified the intent of the package substantially. I would like to step back and offer a brief synthesis, because I think the discussion has surfaced real and useful points, but also some degree of talking past one another.

First, I understand where the concerns about contribution level and general usefulness are coming from. From the outside, KeemenaPreprocessing.jl can appear to sit somewhat awkwardly between existing tokenization packages and more full-featured NLP pipelines, while also introducing its own conventions. Given that framing, questions about overlap, integration, and benchmarks are entirely reasonable.

At the same time, after reading the codebase and the author’s responses, I think it is now clear that the actual contribution is more specific and narrower than that initial impressionand here is a -hopefully useful- summary:

This package is not aiming to be a tokenizer collection or a general-purpose NLP toolkit.

Its core contribution is a fully streaming, corpus-level preprocessing pipeline that produces a deterministic, reproducible, downstream-friendly artifact with predictable memory use.

The multi-layer tokenization with stable cross-references (byte/char/word/sentence/etc.) is a particularly strong and missing feature in the Julia ecosystem. The abstraction level is closer to a preprocessing substrate for modeling pipelines than to a tokenizer or search-engine component.

That said, this distinction is currently clearer in the author’s replies than in the submission artifacts themselves (paper and README). I think several of the reviewer’s concerns stem from that gap rather than from the design itself.

Concretely, I think the submission would benefit from a strengthening of the Statement of Need and the README by explicitly addressing the following two questions:

Who is the intended user? (e.g. researchers preprocessing very large corpora for ML / NLP models)
Who is it not for? (e.g. users looking for a general NLP toolkit)

Responding to these points would help reduce expectations of “feature parity” with other packages.

I agree with the author that loose integration via callables is a reasonable and defensible design choice, especially given:

- the global configuration patterns in packages such as WordTokenizers.jl,
- instability of some upstream APIs and
- the goal of reproducible pipelines.

At the same time, I think the first reviewer’s intuition, that users may assume tighter integration or replacement, is understandable.

A concrete suggestion would be to add a short, explicit section to the docs titled something like:

> Using KeemenaPreprocessing with existing tokenizers

where you show how to use
- WordTokenizers.jl via a callable adapter (as already implemented) (EDIT: the author already added a similar section recently)
- BytePairEncoding.jl as a conceptual example (even if not a hard dependency),

followed by a brief explanation of why hard dependencies are intentionally avoided.

I also sympathize with the concern that performance claims, especially when related to scaling, should be backed by evidence. At the same time, I agree with the author that JOSS does not require exhaustive benchmarking unless strong comparative claims are made.

My suggestion here would be to add one or two focused benchmarks (in README or docs, not necessarily the paper) that would serve as a small demonstration supporting the scalability motivation (e.g. memory behavior of streaming vs non-streaming workflows).

I think the discussion around JLD2 vs mmap conflates two different user expectations. The first is how to exploit portable preprocessing artifacts, while the second concerns high-performance indexed storage. From my reading, KeemenaPreprocessing.jl is clearly targeting the former. I find the choice of JLD2 reasonable in that context, especially since the in-memory bundle is the real interface. That said, the current documentation could better emphasize that JLD2 is a convenience, not a hard constraint, and that the users are free to serialize the bundle differently. Perhaps, a short paragraph in the documentation clarifying why JLD2 was chosen, what trade-offs it implies and how advanced users could plug in alternative storage backends would be helpful.

Finally, I want to highlight that I see the multi-layer, cross-aligned token streams as a genuinely interesting contribution that could support subword tokenization, annotation layers (including LLM-produced spans) and evaluation workflows. This aspect is currently under-emphasized and could be showcased more prominently. A conceptual diagram or worked example showing word tokens + subword tokens, offsets back to raw text and how an external annotation layer could be aligned would make the contribution much clearer to readers unfamiliar with the problem.

Overall, I believe the package is technically solid and conceptually coherent. With improved documentation and positioning many of the concerns raised would largely dissolve and the contribution would be easier to evaluate on its own terms.
