### 2025-12-20 0d9bdf7
now the documentation is improved and benchmarks are included and an example bench for the streaming added in /bench/. The usage of other packages like WordTokenizers.jl and BytePairEncoding.jl is made easier and documented.
