### 2025-12-20 0fe3a61
now the tests are fixed for the new function usage for external tokenizers
