### 2025-12-23 4c05728
CompatHelper: add new compat entry for Revise at version 3, (keep existing compat)
