+++ Project.toml
@@ -1,7 +1,7 @@
 name = "KeemenaPreprocessing"
 uuid = "f4ce45ae-2b88-40b3-ad81-d2d70f7eb3a1"
-authors = ["Alexander V. Mantzaris"]
 version = "0.1.0"
+authors = ["Alexander V. Mantzaris"]
 
 [deps]
 CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
@@ -16,6 +16,7 @@ Unicode = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
 CodecZlib = "0.7"
 Glob = "1"
 JLD2 = "0.5"
+Revise = "3"
 StatsBase = "0.34"
 julia = "1.6"
 
