### 2025-12-25 bbbac9f
CompatHelper: bump compat for JLD2 to 0.6, (keep existing compat)
