### 2025-12-31 #9348 @mantzaris
>     1. I filed a GitHub issue about formatting errors in the documentation. In general you might like to read through it and add check if the structure makes sense, add links between related sections (e.g. tokenizers and custom tokenization) and add signposting sentences since some pages are just headings and examples which doesn't always give enough context to readers.

The issue is addressed and corrected. I have added now cross reference links and added more context.

>     2. What does this mean? "The bundle can be saved or loaded with one function call using the JLD2 format, making it a drop-in dependency for downstream embedding or language-model pipelines inspired by word2vec". How does JLD2 relate to word2vec? If the point is to produce artifacts usable by downstream systems, shouldn't the library output the formats used by those systems. By the way, this video might be somewhat of interest here: https://youtu.be/mtKZbDtZ7PE?si=SgYyXpVwXh8zpybz&t=1604 . I've seen this pattern a few times and haven't heard it captured better than: the "she'll have the steak" explanation.

JLD2 is not related to word2vec. The purpose of that sentence was to say that the PreprocessBundle contains the exact artifacts that word2vec-style training code often needs (a token stream, vocabulary, and related metadata), and that JLD2 is simply a Julia-native persistence approach we can provide for saving and reloading the artifacts of the text processing in a call. For downstream systems outside Julia, the bundle fields are plain arrays and strings (token ids, offsets, vocabulary tables) and can be exported to other formats that a given pipeline expects; adding additional export helpers is a reasonable extension but is orthogonal to the core preprocessing design. The paper's explanation is edited to make this more clear.

>     3. You could replace the "<->" arrows in the paper with a neater:
>          ↔
>         which you can make with `$\leftrightarrow$`

These have been substituted as recommended which do look better now.
