### 2025-12-31 b214cfa
now the paper and documentation are improved according to the reviewer asking for clarity on the paper's mention of JLD2 and for the docs putting in page cross refs and clarifying parts
