### 2026-01-03 mantzaris/KeemenaPreprocessing.jl#8 @frankier
preprocess_corpus is not done when it returns
If I run a script like this:

```
using KeemenaPreprocessing

files = ["11-0.txt", "35-0.txt"]
cfg   = PreprocessConfiguration(minimum_token_frequency = 5)
bund1 = preprocess_corpus(files; config  = cfg, save_to = "corpus.jld2")

@info "Done"
```

It takes a long time to terminate after printing "Done". I assume this is because `preprocess_corpus(...)` is still running.

This is surprising and could cause quite a bit of confusion. e.g. This example will probably not work if run as written:

https://mantzaris.github.io/KeemenaPreprocessing.jl/dev/guides/quickstart/#Saving-and-loading-bundles

Since the saving is not finished before the loading starts.
