### 2026-01-03 #9348 @frankier
> JLD2 is not related to word2vec. The purpose of that sentence was to say that the PreprocessBundle contains the exact artifacts that word2vec-style training code often needs (a token stream, vocabulary, and related metadata), and that JLD2 is simply a Julia-native persistence approach we can provide for saving and reloading the artifacts of the text processing in a call. For downstream systems outside Julia, the bundle fields are plain arrays and strings (token ids, offsets, vocabulary tables) and can be exported to other formats that a given pipeline expects; adding additional export helpers is a reasonable extension but is orthogonal to the core preprocessing design. The paper's explanation is edited to make this more clear.

Thanks for clarifying. However, I'm still a bit confused. Here's what you seem to be outlining:

```
KeemenaPreprocessing -> JLD2 -> Word2Vec (not the original program -- either a reimplementation or wrapper in Julia)
```

The text seems to be recommending the JLD2 format again for large scale work as a solution above and beyond the support in the library (simply implementing it but not mentioning it here could be interpreted as a feature for small scale work only). Using Word2Vec as an example hints towards this since it is an online algorithm rather than full batch. If it were full batch it wouldn't really matter since everything would have to be loaded into memory anyway. I have two points I would like to bring in here: a more specific one and a more general one.

Specific point: You need to do at least two passes of the corpus for your Word2Vec (at least typically). The first pass to build up frequencies and so build your vocabulary and the second pass to encode the documents. However, you can run Word2Vec training already in the second pass. It seems unnecessary to make a third pass. In fact is is possible in these cases and somewhat common to take the first pass with a smaller (but still large) corpus or corpus sample to avoid making multiple passes over truly giant corpora. It's a bit confusing that you talk about JLD2, since you seem to have the possibility of doing two passes here: https://mantzaris.github.io/KeemenaPreprocessing.jl/dev/guides/streaming/ . I suppose if you were training multiple models this might seem a bit wasteful? What about saving only the results of the first pass (i.e. the full and/or truncated vocabulary) and iterating over the corpora from the text files in the second? This would also allow "training" you vocabulary model and your embedding model on different corpora. I think this is how things are done in gensim for example. In fact, you could do a lot worst than copying parts of the architecture of gensim for this package.

General point: Can you really recommend this format to your users for working with large corpora? It is based on HDF5, which is a general purpose and somewhat crufty format, but there are plenty of simpler options that are likely to cause less headaches. As an indicative example put the following in `mytest.jl`:

```julia
using KeemenaPreprocessing

files = ["11-0.txt", "35-0.txt"]
cfg   = PreprocessConfiguration(minimum_token_frequency = 5)
bund1 = preprocess_corpus(files; config  = cfg, save_to = "corpus.jld2")

@info "Sizes" sum((filesize(f) for f in files)) filesize("corpus.jld2")
```

And run it from your project root like so:

```bash
wget https://www.gutenberg.org/files/35/35-0.txt && wget https://ww
w.gutenberg.org/files/11/11-0.txt && julia --project=. mytest.jl
```

Then you'll get:

```
┌ Info: Sizes
│   sum((filesize(f) for f = files)) = 335938
└   filesize("corpus.jld2") = 1076698
```

So this increases the (uncompressed) corpus size by 3X. This is presumably the last thing you would want if you had a very large corpus.

As far as I understand there's no real native support for sharding in HDF5. This is also odd since there is some kind of sharding support in your library. Ideally the abstractions offered by the library would steer the user towards a performant implementation.

By the way, this sentence is a sort of "hint" which gave me a vibe that this submission is premature. This package implements the first part of a pipeline. Any claims about scalability would be more convincing if given in terms of a full pipeline. I think if you implemented a pipeline involving word2vec (or similar) yourself, you would probably come to similar conclusions about the problems of dumping the whole corpus as JLD2 as an intermediate step. By advertising this software via this publication already, it seems you are inviting prospective users to fall into this trap. Ideally I would like you to carry on and implement this full pipeline, and then publish the version of this package that has been refactored such that it can accommodate it rather than attempting to sort of rebuttal and comb-over or reducing the claims about the package. For this reason, I sort of miss the possibility in this review process to make a clear substantial revisions/constructive reject (but come back later) decision. @lockwo Maybe you can advise here?

P.S. I filed another issue https://github.com/mantzaris/KeemenaPreprocessing.jl/issues/8
