### 2026-01-03 mantzaris/KeemenaPreprocessing.jl#8 @mantzaris
Thanks for the report. I tried to reproduce both concerns:

preprocess_corpus(...; save_to="...") returning before the .jld2 file is fully written (which would make the "save then load" quickstart snippet unsafe).

Julia taking a long time to terminate after printing Done.

On my machine with current code, I cannot reproduce either. The .jld2 file is readable immediately after preprocess_corpus returns, the loaded bundle matches the in-memory bundle, and the measured "post-Done" exit delay is about 0.03 seconds.

Repro script (downloads 2 Gutenberg files, saves, immediately loads, compares token ids, measures exit delay)
```julia
using KeemenaPreprocessing
using Downloads

function download_if_missing(url::String, path::String)
    if !isfile(path)
        Downloads.download(url, path)
    end
    return path
end

files = [
    download_if_missing("https://www.gutenberg.org/files/11/11-0.txt", "11-0.txt"),
    download_if_missing("https://www.gutenberg.org/files/35/35-0.txt", "35-0.txt"),
]

configuration = PreprocessConfiguration(minimum_token_frequency = 5)
save_path = "corpus.jld2"

done_time = Ref{Float64}(NaN)
atexit(() -> begin
    if !isnan(done_time[])
        @info "Exit delay after printing Done (seconds)" delay = (time() - done_time[])
    end
end)

@info "Step 1: preprocess_corpus(save_to=...)" save_path = save_path
bund1 = preprocess_corpus(files; config = configuration, save_to = save_path)

# Quick sanity check: file is stable right after return
size1 = stat(save_path).size
sleep(0.5)
size2 = stat(save_path).size
@info "File size stable after return?" size1 = size1 size2 = size2 stable = (size1 == size2)

@info "Step 2: immediate load"
bund2 = load_preprocess_bundle(save_path)

tokens1 = get_token_ids(bund1, :word)
tokens2 = get_token_ids(bund2, :word)
@info "Token ids match?" n1 = length(tokens1) n2 = length(tokens2) equal = (tokens1 == tokens2)

done_time[] = time()
@info "Done"
```
and example output:
```
me@me:~/Downloads$ time julia --project=. testo2.jl
┌ Info: Step 1: preprocess_corpus(save_to=...)
└   save_path = "corpus.jld2"
┌ Info: File size stable after return?
│   size1 = 1076698
│   size2 = 1076698
└   stable = true
[ Info: Step 2: immediate load
┌ Info: Token ids match?
│   n1 = 53797
│   n2 = 53797
└   equal = true
[ Info: Done
┌ Info: Exit delay after printing Done (seconds)
└   delay = 0.021934032440185547

real	0m6.388s
user	0m5.811s
sys	0m0.512s
```
So, at least in this setup, save_to behaves synchronously and the quickstart "save then load" pattern is safe.

you can also run this as a script to see the difference between slow start up and downloads and compilation from the slow exit:
```julia
using KeemenaPreprocessing
using Downloads
using Dates

atexit(() -> @info "atexit reached" timestamp=string(now()))

files = ["11-0.txt", "35-0.txt"]
Downloads.download("https://www.gutenberg.org/files/11/11-0.txt", files[1])
Downloads.download("https://www.gutenberg.org/files/35/35-0.txt", files[2])

cfg = PreprocessConfiguration(minimum_token_frequency = 5)
bund1 = preprocess_corpus(files; config = cfg, save_to = "corpus.jld2")
@info "Done" timestamp=string(now())
```
