### 2026-01-03 3705685
now a documentation edit is included and a change to the Project.toml
