### 2026-01-03 #9348 @mantzaris
@frankier, I think there are two separate points here:

1. Are we forcing an unnecessary extra pass for Word2Vec-style training?
2. Is it appropriate to recommend `save_to = "corpus.jld2"` (JLD2) as a default strategy for very large corpora?

KeemenaPreprocessing’s streaming preprocessing is already explicitly **two-pass**:

- Pass 1 (vocabulary pass): constant-memory scan to count tokens and build the `Vocabulary`
  - This pass is skipped if you supply an existing `vocab = ...`
- Pass 2 (chunk pass): encode the corpus into fixed-size `PreprocessBundle` chunks (approx `chunk_tokens`) and yield them

Because pass 2 yields bundles via a Channel, an online algorithm (eg, Word2Vec/SGNS) can train directly inside the chunk loop if desired. That means there is no need to write an intermediate encoded corpus and then read it back (no inherent third pass):

a general approach can be (vocab on one corpus/sample, train on another):
```julia
using KeemenaPreprocessing

cfg = PreprocessConfiguration(minimum_token_frequency = 5, tokenizer_name = :whitespace)

for bund in preprocess_corpus_streaming(files; cfg, chunk_tokens = 250_000)
    word_ids = get_token_ids(bund, :word)

    # optional: recover document/sentence boundaries from offsets if needed
    # wc = get_corpus(bund, :word)
    # wc.document_offsets / wc.sentence_offsets / ....

    update_model!(model, word_ids)# your training step
end
```

This is another approach that works: build the vocabulary once (possibly on a sample corpus), then pass it into streaming preprocessing to skip a pass on the full corpus:
```
sample_bundle = preprocess_corpus(sample_files; config = cfg)
vocab = sample_bundle.levels[:word].vocabulary

for bund in preprocess_corpus_streaming(full_files; cfg, vocab = vocab, chunk_tokens = 250_000)
    update_model!(model, get_token_ids(bund, :word))
end
```

So the architecture is not like that comment: KeemenaPreprocessing -> JLD2 -> Word2Vec.
The intended large-corpus path is KeemenaPreprocessing streaming -> training loop, with JLD2 as an optional persistence convenience which can be used for Word2Vec later on or other models.

Docs reference for this: the Streaming Pipeline guide:
https://mantzaris.github.io/KeemenaPreprocessing.jl/dev/guides/streaming

Also, the package is not locked into JLD2: PreprocessBundle is a plain Julia object, and users can store large arrays with a different backend if they need memory mapping, indexed random access, or cross-language IO. This is now stated explicitly in the Quick Start:
https://mantzaris.github.io/KeemenaPreprocessing.jl/dev/guides/quickstart/#Why-JLD2-(and-when-to-use-something-else)

About the reported 3x on-disk size increase:
It seems like your .jld2 output is not a compressed version of the raw text which is an option to choose. Since itt stores additional derived structures (integer token ids, offset arrays, vocabulary tables, plus metadata, and the cleaned raw text for alignment support). It is therefore expected that it can be larger than the raw corpus in some cases. But the artifacts will take much less space than the original text depending if the full text is stored with it or not. That trade-off (extra structure vs raw bytes) is exactly why the streaming interface exists for large-scale work.

If the goal is smaller artifacts, JLD2 itself supports dataset compression (and other backends can be used). save_to is convenience caching, and streaming is the intended large-corpus workflow. If users do use persistance, they should use the  'compression' of JLD2 which is provided (JLD2 compression docs: https://juliaio.github.io/JLD2.jl/stable/compression/). JLD2 is widely supported and integrates cleanly with Julia workflows.

In regards to 'sharding': the library's sharding is at the preprocessing level: preprocess_corpus_streaming and preprocess_corpus_streaming_chunks yield a sequence of PreprocessBundle chunks. If persistence is needed, users can simply save one file per chunk (application-level sharding). This avoids a monolithic intermediate file and matches downstream training patterns.

Related: issue #8 (async save / delayed exit):
PS I answered that 'issue' you posted with some minimal code that uses data from Gutenberg text files, I cannot reproduce an async save: preprocess_corpus(...; save_to=...) returns after the file is stable, and an immediate load_preprocess_bundle succeeds with identical token ids. (it may be Julia shutdown / GC / IO flushing rather than asynchronous saving) But running that code test shows that it passes. If you run that code you will see it working.
