### 2026-01-04 #9348 @frankier
If you've looked into the issue and it doesn't replicate that's good enough for me. You're right that it was a bit of a sloppy issue based on something I ran into while writing the example code. It could be like you say -- or maybe it's something to do with a package loaded in my startup.jl.

In general I think there is more that could be done to represent your software well and guide users towards good patterns of usage. Whether this is will actually be enough of a problem that it will prevent people reusing the package (included in one of the checklist criteria) is something I don't feel like I can judge well any more (keeping in mind that in many cases, the specific code this library would replace is a few 10s of lines).

@lockwo Can I please withdraw as reviewer for this submission?
