### 2026-01-07 #9348 @lockwo
Hi @mantzaris, just a heads up, but I will be inviting another reviewer (JOSS requires 2 reviews). All this means is I will find a reviewer (similar to the pre review stage) then add them to this review (should all work within the current issue), then the process can proceed. I recognize there has been a lot of discussion so far, and I appreciate the reviewers and authors engaging meaningfully about this package.
