### 2026-01-13 #9348 @lockwo
@akki2825 Thanks for agreeing to review this submission!

I just want to give you a heads up that this is an ongoing review, so you'll notice a lot of discussion already happened (no worries joining in, the timeline adapts to new reviewers). Another heads up, the generate checklist function might generate the new checklist for the new scope (we are undergoing a lot of changes as you can see here https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future, but this falls under the old scope, so if the checklist looks different, you can just copy the one from here https://github.com/openjournals/joss-reviews/issues/9348#issuecomment-3538671943)

As you can see above, you each should use the command @editorialbot generate my checklist to create your review checklist. @editorialbot commands need to be the first thing in a new comment.

As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines here: https://joss.readthedocs.io/en/pre-2026-submission/reviewer_guidelines.html.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention https://github.com/openjournals/joss-reviews/issues/9348 so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 3-4 weeks [as discussed over email, @atantos, I know you are unable to start this until December, so I will ping you then]. Please let me know if either of you require some more time. We can also use editorialbot (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Note that the main/top issue text also contains information, such as how to generate your checklists.
