### 2026-01-18 5de4010
CompatHelper: bump compat for JLD2 to 0.6, (keep existing compat)
