### 2026-01-27 #9348 @akki2825
I've checked off most of my checklist. Two items remain unchecked.

Community guidelines: There is no contributing file or code of conduct or any guidance in the README.

References: There are two missing DOIs. In line 46, the same citation appears twice in the same sentence. The Mikolov 2013 paper is missing co-authors, and the Gensim author name is misspelled.

Few more minor changes:

- Remove the commented code in docs/make.jl
- In `src/processing/vocabulary.jl` line 127, `$` is missing in the error message.
