### 2026-01-27 ced0a55
Fixing the paper citations, adding community guidelines and the contributing notes. Another fix to an error message in vocabulary.jl
