+++ paper/paper.bib
@@ -16,14 +16,16 @@
   title={spaCy: Industrial-strength natural language processing in python},
   author={Honnibal, Matthew and Montani, Ines and Van Landeghem, Sofie and Boyd, Adriane and others},
   year={2020},
-  publisher={Zenodo, Honolulu, HI, USA}
+  publisher={Zenodo, Honolulu, HI, USA},
+  doi={10.5281/zenodo.1212303}
 }
 
-@article{vrehuuvrek2010software,
+@inproceedings{vrehuuvrek2010software,
   title={Software framework for topic modelling with large corpora},
-  author={Rehrek, Radim and Sojka, Petr},
+  author={Řehůřek, Radim and Sojka, Petr},
   year={2010},
-  publisher={University of Malta}
+  publisher={University of Malta},
+  doi={10.13140/2.1.2393.1847}
 }
 
 @article{kaushal2020wordtokenizers,
@@ -33,22 +35,26 @@
   volume={5},
   number={46},
   pages={1956},
-  year={2020}
+  year={2020},
+  doi={10.21105/joss.01956}
 }
 
 @book{bird2009natural,
   title={Natural language processing with Python: analyzing text with the natural language toolkit},
   author={Bird, Steven and Klein, Ewan and Loper, Edward},
   year={2009},
-  publisher={" O'Reilly Media, Inc."}
+  publisher={" O'Reilly Media, Inc."},
+  doi={10.5555/1717171}
 }
 
 @article{mikolov2013efficient,
   title={Efficient estimation of word representations in vector space},
-  author={Mikolov, Tomas},
+  author={Mikolov, Tomas and Chen, Kai and Corrado, Greg and Dean, Jeffrey},
   journal={arXiv preprint arXiv:1301.3781},
   volume={3781},
-  year={2013}
+  year={2013},
+  doi={10.48550/arXiv.1301.3781},
+  url={https://doi.org/10.48550/arXiv.1301.3781}
 }
 
 @inproceedings{manning2014stanford,
@@ -56,5 +62,7 @@
   author={Manning, Christopher D and Surdeanu, Mihai and Bauer, John and Finkel, Jenny Rose and Bethard, Steven and McClosky, David},
   booktitle={Proceedings of 52nd annual meeting of the association for computational linguistics: system demonstrations},
   pages={55--60},
-  year={2014}
+  year={2014},
+  doi={10.3115/v1/P14-5010},
+  url={https://doi.org/10.3115/v1/P14-5010}
 }
