### 2026-01-27 #9348 @mantzaris
@akki2825
- Community guidelines and code of conduct have been added to the README and the docs
- I fixed the repeated citation in the same sentence
- I found and added the missing DOIs
- The names for the Gensim are now correctly spelled (as I can tell)
- The commented code is removed
- The error message in vocabulary.jl is fixed

The edits have been pushed. Let me know if there is any change you'd like to see to the Readme or docs related to the community guidelines etc.
