+++ paper/paper.bib
@@ -12,12 +12,13 @@
   journal   = {{SIAM} Review}
 }
 
-@article{honnibal2020spacy,
+@software{honnibal2020spacy,
   title={spaCy: Industrial-strength natural language processing in python},
   author={Honnibal, Matthew and Montani, Ines and Van Landeghem, Sofie and Boyd, Adriane and others},
   year={2020},
   publisher={Zenodo, Honolulu, HI, USA},
-  doi={10.5281/zenodo.1212303}
+  doi={10.5281/zenodo.1212303},
+  url={https://doi.org/10.5281/zenodo.1212303}
 }
 
 @inproceedings{vrehuuvrek2010software,
@@ -25,7 +26,10 @@
   author={Řehůřek, Radim and Sojka, Petr},
   year={2010},
   publisher={University of Malta},
-  doi={10.13140/2.1.2393.1847}
+  doi={10.13140/2.1.2393.1847},
+  publisher={ELRA},
+  url={http://is.muni.cz/publication/884893/en},
+  booktitle={Proceedings of the {LREC} 2010 Workshop on New Challenges for {NLP} Frameworks}
 }
 
 @article{kaushal2020wordtokenizers,
@@ -36,7 +40,8 @@
   number={46},
   pages={1956},
   year={2020},
-  doi={10.21105/joss.01956}
+  doi={10.21105/joss.01956},
+  url={https://doi.org/10.21105/joss.01956}
 }
 
 @book{bird2009natural,
