### 2026-01-28 #9348 @mantzaris
@akki2825 I looked through them and found missing fields and added them to those references and pushed that revised version. Does it all look ok now?
