+++ paper/paper.bib
@@ -12,13 +12,12 @@
   journal   = {{SIAM} Review}
 }
 
-@software{honnibal2020spacy,
-  title={spaCy: Industrial-strength natural language processing in python},
-  author={Honnibal, Matthew and Montani, Ines and Van Landeghem, Sofie and Boyd, Adriane and others},
-  year={2020},
-  publisher={Zenodo, Honolulu, HI, USA},
-  doi={10.5281/zenodo.1212303},
-  url={https://doi.org/10.5281/zenodo.1212303}
+
+@article{Honnibal_spaCy_Industrial-strength_Natural_2020,
+author = {Honnibal, Matthew and Montani, Ines and Van Landeghem, Sofie and Boyd, Adriane},
+doi = {10.5281/zenodo.1212303},
+title = {{spaCy: Industrial-strength Natural Language Processing in Python}},
+year = {2020}
 }
 
 @inproceedings{vrehuuvrek2010software,
