+++ paper/paper.bib
@@ -48,7 +48,9 @@ year = {2020}
   author={Bird, Steven and Klein, Ewan and Loper, Edward},
   year={2009},
   publisher={" O'Reilly Media, Inc."},
-  doi={10.5555/1717171}
+  isbn      = {978-0-596-51649-9},
+  url       = {https://dl.acm.org/doi/10.5555/1717171},
+  doi       = {10.5555/1717171}  
 }
 
 @article{mikolov2013efficient,
