### 2026-01-31 #9348 @mantzaris
@lockwo

- Figshare DOI: 10.6084/m9.figshare.31216063

- Regarding the missing DOI for this reference:
@book{bird2009natural,
  title={Natural language processing with Python: analyzing text with the natural language toolkit},
  author={Bird, Steven and Klein, Ewan and Loper, Edward},
  year={2009},
  publisher={" O'Reilly Media, Inc."},
  isbn      = {978-0-596-51649-9},
  url       = {https://dl.acm.org/doi/10.5555/1717171},
  doi       = {10.5555/1717171}
}

I cannot find a better DOI but this should suffice in practice. All the information exists where it is pointed to by the reference
