### 2026-01-31 #9348 @lockwo
Hmm I see, maybe just delete the DOI field and leave the ACM url?

Other than that a few points:
- `WordTokenizers. Jl` isn't formatted right in the bib (I think should be {WordTokenizers.jl})
- Extra space in front of O'Reilly
- Inline packages (e.g. x.jl) have ` around them
