+++ paper/paper.bib
@@ -32,7 +32,7 @@ year = {2020}
 }
 
 @article{kaushal2020wordtokenizers,
-  title={WordTokenizers. jl: Basic tools for tokenizing natural language in Julia},
+  title={{WordTokenizers.jl}: Basic tools for tokenizing natural language in Julia},
   author={Kaushal, Ayush and White, Lyndon and Innes, Mike and Kumar, Rohit},
   journal={Journal of Open Source Software},
   volume={5},
@@ -47,10 +47,9 @@ year = {2020}
   title={Natural language processing with Python: analyzing text with the natural language toolkit},
   author={Bird, Steven and Klein, Ewan and Loper, Edward},
   year={2009},
-  publisher={" O'Reilly Media, Inc."},
+  publisher={"O'Reilly Media, Inc."},
   isbn      = {978-0-596-51649-9},
-  url       = {https://dl.acm.org/doi/10.5555/1717171},
-  doi       = {10.5555/1717171}  
+  url       = {https://dl.acm.org/doi/10.5555/1717171}
 }
 
 @article{mikolov2013efficient,
