### 2026-02-02 #9348 @mantzaris
> Did the new generation not work? I see the stuff like WordTokenizers. Jl as the same?

@lockwo I compiled the .bib using this:
```
@article{kaushal2020wordtokenizers,
  title={{WordTokenizers.jl}: Basic tools for tokenizing natural language in Julia},
  author={Kaushal, Ayush and White, Lyndon and Innes, Mike and Kumar, Rohit},
  journal={Journal of Open Source Software},
  volume={5},
  number={46},
  pages={1956},
  year={2020},
  doi={10.21105/joss.01956},
  url={https://doi.org/10.21105/joss.01956}
}
```

I pushed this after a new compilation just to be sure and it seems to be correct.
