### 2026-02-03 #9348 @lockwo
Great, now at this point could you:

Also one other thing on the paper (this can be done before or after, changing the paper doesn't require changing the above):

- Basic tools for tokenizing natural language in julia (J in Julia should be upper case)
