### 2026-02-03 #9348 @mantzaris
@lockwo

- release (v0.1.1): https://github.com/mantzaris/KeemenaPreprocessing.jl/releases/tag/v0.1.1
- Figshare DOI: 10.6084/m9.figshare.31216063

Does this cover it?
