+++ paper/paper.bib
@@ -32,7 +32,7 @@ year = {2020}
 }
 
 @article{kaushal2020wordtokenizers,
-  title={{WordTokenizers.jl}: Basic tools for tokenizing natural language in Julia},
+  title={{WordTokenizers.jl}: Basic tools for tokenizing natural language in {J}ulia},
   author={Kaushal, Ayush and White, Lyndon and Innes, Mike and Kumar, Rohit},
   journal={Journal of Open Source Software},
   volume={5},
