### 2026-02-04 4bd2c21
Change date format from 'July 7 2025' to '7 July 2025'
