### 2026-02-04 #9348 @lockwo
@mantzaris Can you make the metadata of the archive title be the same as the paper? (KeemenaPreprocessing.jl: a text pre-processor designed for extensibility and compatibility for modern day text analysis -> KeemenaPreprocessing.jl: Unicode-Robust Cleaning, Multi-Level Tokenisation & Streaming Offset Bundling for Julia NLP). In zenodo this doesn't require a new archive link so I assume one can edit the meta data similarly in fig share.

(also I put some minor stuff in https://github.com/mantzaris/KeemenaPreprocessing.jl/pull/9)
