### 2026-02-05 #9348 @lockwo
Did it generate a new link? If I go to https://doi.org/10.6084/m9.figshare.31216063 I see the same title
