### 2026-02-06 7b359f3
now the documentation has examples of how to use a python tokenizer and also using Python's spaCy as well
