### 2026-02-06 #9348 @lockwo
@mantzaris  - please let me know if you see any issues with the proof that this generates
