### 2026-02-06 #9348 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
