### 2026-02-06 #9348 @lockwo
Btw for @openjournals/dsais-eics, the skip DOI one seemed to have a known broken prefix, so to avoid the error, the DOI was moved from the DOI field to the URL field (and the URL/DOI is correct), as recommended here https://github.com/openjournals/joss-reviews/issues/9348#issuecomment-3825433485.
