### 2026-02-06 #9348 @lockwo
Ah one last thing I just noticed with the proof, the Python in the "Natural language processing with python" ref should be with capital P for Python.
