### 2026-02-06 #9348 @mantzaris
> Ah one last thing I just noticed with the proof, the Python in the "Natural language processing with python" ref should be with capital P for Python.

@lockwo  , I made the bib entry:

@book{bird2009natural,
  title={Natural language processing with {Python}: analyzing text with the natural language toolkit},
  author={Bird, Steven and Klein, Ewan and Loper, Edward},
  year={2009},
  publisher={"O'Reilly Media, Inc."},
  isbn      = {978-0-596-51649-9},
  url       = {https://dl.acm.org/doi/10.5555/1717171}
}

which should fix that capitalization. Does everything look ok now?
