### 2026-02-13 #9348 @lockwo
Yes, I think it should be good, the track editor in chief just has to do the final step. I'll regenerate the pdf since it was changed
