### 2026-02-22 e502a66
CompatHelper: bump compat for JLD2 to 0.6, (keep existing compat)
