### 2026-02-23 #9348 @crvernon
@mantzaris could you replace the "&" in your title with "and" please?  The special character is causing issues with the formatting of the title in JOSS.   Let me know when this is done and I'll recompile.  Thanks.
